refactor: separate TodoInput as a component
